Updated functional tests to use pytest, using .env config
