+++ tests/functional_tests/basic_workflow_test.py
@@ -18,33 +18,32 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
+import covalent as ct
+import pytest
 
-import json
-import os
-import subprocess
-import sys
-
-TERRAFORM_OUTPUTS = {}
-
-try:
-    terraform_dir = os.getenv("TF_DIR")
-    proc = subprocess.run(
-        [
-            "terraform",
-            f"-chdir={terraform_dir}",
-            "output",
-            "-json",
-        ],
-        check=True,
-        capture_output=True,
-    )
-    TERRAFORM_OUTPUTS = json.loads(proc.stdout.decode())
-except Exception as e:
-    pass
+from tests.functional_tests.fixtures.executor import executor
 
 
-def get(key: str, default):
-    try:
-        return TERRAFORM_OUTPUTS[key]["value"]
-    except KeyError:
-        return default
+@pytest.mark.functional_tests
+def test_basic_workflow():
+    @ct.electron(executor=executor)
+    def join_words(a, b):
+        return ", ".join([a, b])
+
+    @ct.electron
+    def excitement(a):
+        return f"{a}!"
+
+    @ct.lattice
+    def basic_workflow(a, b):
+        phrase = join_words(a, b)
+        return excitement(phrase)
+
+    # Dispatch the workflow
+    dispatch_id = ct.dispatch(basic_workflow)("Hello", "World")
+    result = ct.get_result(dispatch_id=dispatch_id, wait=True)
+    status = str(result.status)
+
+    print(result)
+
+    assert status == str(ct.status.COMPLETED)
